+++ packages/tree-view/src/TreeView.js
@@ -108,7 +108,7 @@ TreeView.propTypes = {
       id: PropTypes.number,
       parentId: PropTypes.number,
       meta: PropTypes.shape({
-        label: PropTypes.string,
+        label: PropTypes.node,
         collapsed: PropTypes.bool,
         active: PropTypes.bool,
         icon: PropTypes.node