+++ packages/tree-view/src/behaviors/TreeViewBehavior.js
@@ -194,7 +194,7 @@ TreeViewBehavior.propTypes = {
       id: PropTypes.number,
       parentId: PropTypes.number,
       meta: PropTypes.shape({
-        label: PropTypes.string,
+        label: PropTypes.node,
         collapsed: PropTypes.bool,
         active: PropTypes.bool,
         icon: PropTypes.node